requirements.txt
# Requirements for the Selektaaa project
librosa
numpy
pandas
scikit-learn
torch
tqdm
pyyaml
